Fix test
didn't mean to blow up on invocation, but on promise resolve. The nodeService key is a hook that needs to be truthy to prevent a fetch, but also in these cases, it needs to reject a promise to let the developer know they did something unexpected on the node-side of code execution.